refactor: relocate FallbackPolicy* and QueryHandlerAttribute* to mirror Brighter (#321)
Tidy First (structural-only): move `QueryHandlerAttribute*` to root namespace
`Paramore.Darker`, and `FallbackPolicy*` (attributes + decorators) into
`Policies/{Attributes,Handlers}/` feature folders. Update production-code +
side-package `using`s for the namespace moves. Test-file usings are swept
in Step 5.
