+++ src/Paramore.Darker.QueryLogging/QueryLoggingAttributeAsync.cs
@@ -1,5 +1,4 @@
 using System;
-using Paramore.Darker.Attributes;
 
 namespace Paramore.Darker.QueryLogging
 {
