+++ src/Paramore.Darker.QueryLogging/QueryLoggingDecoratorAsync.cs
@@ -4,7 +4,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
-using Paramore.Darker.Decorators;
+using Paramore.Darker.Policies.Handlers;
 using Paramore.Darker.Exceptions;
 using Paramore.Darker.Logging;
