+++ src/Paramore.Darker/PipelineBuilder.cs
@@ -5,7 +5,6 @@ using System.Reflection;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using Paramore.Darker.Attributes;
 using Paramore.Darker.Exceptions;
 using Paramore.Darker.Logging;
 using System.Runtime.ExceptionServices;
