+++ src/Paramore.Darker/Policies/Handlers/FallbackPolicyDecorator.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.Extensions.Logging;
 
-namespace Paramore.Darker.Decorators
+namespace Paramore.Darker.Policies.Handlers
 {
     public class FallbackPolicyDecorator<TQuery, TResult> : IQueryHandlerDecorator<TQuery, TResult>
         where TQuery : IQuery<TResult>
